refactor: split server integration tests
