+++ apps/server/src/test-helpers.ts
@@ -1,0 +1,7 @@
+import type { LiveKitConfig } from "./livekit.js";
+
+export const TEST_LIVEKIT_CONFIG: LiveKitConfig = {
+  url: "ws://localhost:7880",
+  apiKey: "devkey",
+  apiSecret: "secret",
+};
